refactor: modularize terminal window
