+++ components/terminal/terminal-window/index.ts
@@ -1,0 +1,5 @@
+export { TerminalWindow } from './terminal-window'
+export type { TerminalCommand } from './types'
+export { CommandHistory } from './command-history'
+export { AISuggestionsPanel } from './ai-suggestions-panel'
+export { PerformanceOverlay } from './performance-overlay'
